进一步完善移位的提示

+++ src/utils/ExpressionUtil.java
@@ -22,7 +22,7 @@ public class ExpressionUtil {
         }
         return result;
     }
-    public static int calShiftSequence(int init,int num){
+    public static long calShiftSequence(long init,long num){
         if(num>0){//左移
             return init<<num;
         }else {//右移
